task_1_transactions/requirements.txt
greenlet==3.4.0
psycopg2-binary==2.9.11
python-dotenv==1.2.2
SQLAlchemy==2.0.23
typing_extensions==4.15.0
